add card to 10 character in name

+++ app/Http/Controllers/Admin/PenemuanController.php
@@ -40,6 +40,9 @@ class PenemuanController extends Controller
 
 public function destroy($id)
 {
+     if (Auth::user()->hasRole('berwenang')) {
+        abort(403, 'Anda tidak memiliki izin untuk menghapus data.');
+    }
     $pengumuman = Pengumuman::penemuan()->findOrFail($id);
     $pengumuman->delete();
